requirements.txt
bcrypt==4.1.2
Flask==3.0.3
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.32.3
